scripts/docker_configurator/requirements.txt
mako
pyyaml
pyinstaller
